requirements.txt
pandas>=2.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
python-magic>=0.4.27
html2text>=2020.1.16
fastapi>=0.100.0
uvicorn>=0.20.0
psycopg[binary]>=3.1.12
psycopg-pool>=3.1.1
psycopg2>=2.9.0
supabase>=2.0.0
backoff>=2.2.1
python-dotenv>=1.0.0
pydantic>=2.0.0
postgrest>=0.10.6
pyshorteners>=1.0.1
slowapi>=0.1.9
# Authentication providers
httpx>=0.24.0  # For Auth0 and Clerk HTTP requests
PyJWT>=2.8.0   # For JWT token decoding and verification
python-jose[cryptography]>=3.3.0  # For advanced JWT operations
passlib[bcrypt]>=1.7.4  # For password hashing (if implementing custom auth)
alembic>=1.13.0
